server/requirements.txt
alabaster==1.0.0
argcomplete==3.6.2
attrs==25.4.0
Authlib==1.6.5
babel==2.17.0
bandit==1.8.6
black==25.11.0
blinker==1.9.0
build==1.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cryptography==46.0.3
dnspython==2.8.0
docutils==0.21.2
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meson==1.9.1
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pbr==7.0.3
pillow==12.0.0
pip-tools==7.5.2
platformdirs==4.5.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
roman-numerals-py==3.1.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stevedore==5.4.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1+cpu
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
